requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysqlclient
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
PyMySQL
